pyproject.toml
[project]
name = "brainstorm-track2"
version = "0.1.0"
description = "BrainStorm 2026 Track 2 - Neural Data Viewer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "huggingface-hub>=0.20.0",
    "pynput>=1.7.0",
    "pandas-stubs~=2.3.3",
    "orjson>=3.9.0",
    "scipy-stubs>=1.15.3.0",
]

[project.scripts]
brainstorm-stream = "scripts.stream_data:app"
brainstorm-serve = "scripts.serve:cli"
brainstorm-backend = "scripts.backend:cli"
brainstorm-control = "scripts.control_client:app"
brainstorm-all = "scripts.launcher:main"

[build-system]
requires = ["hatchling", "setuptools"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts", "web_viewer"]

[tool.hatch.build]
include = [
    "scripts/**/*.py",
    "web_viewer/**/*",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "TC",   # flake8-type-checking (move type-only imports behind TYPE_CHECKING)
    "RUF",  # Ruff-specific rules
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ty.environment]
python-version = "3.10"

[tool.uv.extra-build-dependencies]
evdev = ["setuptools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = ["e2e: end-to-end tests requiring a running server (use --base-url)"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-playwright>=0.7.2",
    "ty>=0.0.13",
]
